+++ lib/src/components/message/view.dart
@@ -1,19 +1,26 @@
 // Flutter imports:
+import 'dart:async';
+
+import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/scheduler.dart';
 
 // Package imports:
 import 'package:zego_uikit/zego_uikit.dart';
 
 // Project imports:
 import 'package:zego_uikit_prebuilt_live_streaming/src/config.dart';
+import 'package:zego_uikit_prebuilt_live_streaming/src/inner_text.dart';
 
 /// @nodoc
 class ZegoInRoomLiveMessageView extends StatefulWidget {
   final ZegoInRoomMessageConfig? config;
   final ZegoAvatarBuilder? avatarBuilder;
+  final ZegoInnerText innerText;
 
   const ZegoInRoomLiveMessageView({
     Key? key,
+    required this.innerText,
     this.config,
     this.avatarBuilder,
   }) : super(key: key);
@@ -45,6 +52,12 @@ class _ZegoInRoomLiveMessageViewState extends State<ZegoInRoomLiveMessageView> {
                 (BuildContext context, ZegoInRoomMessage message, _) {
                   return ZegoInRoomMessageViewItem(
                     message: message,
+                    avatarLeadingBuilder: widget.config?.avatarLeadingBuilder,
+                    avatarTailingBuilder: widget.config?.avatarTailingBuilder,
+                    nameLeadingBuilder: widget.config?.nameLeadingBuilder,
+                    nameTailingBuilder: widget.config?.nameTailingBuilder,
+                    textLeadingBuilder: widget.config?.textLeadingBuilder,
+                    textTailingBuilder: widget.config?.textTailingBuilder,
                     avatarBuilder: widget.avatarBuilder,
                     showName: widget.config?.showName ?? true,
                     showAvatar: widget.config?.showAvatar ?? true,